Reverted last commit.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@6496 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ sound/LADSPAPluginInstance.cpp
@@ -179,7 +179,7 @@ LADSPAPluginInstance::getLatency()
 {
     if (m_latencyPort) {
 	if (!m_run) run(RealTime::zeroTime);
-	return static_cast<unsigned int>(*m_latencyPort);
+	return *m_latencyPort;
     }
     return 0;
 }
